Move error to new file
